Тот момент, когда смотрится красиво, но сокет сам собой где-то закрывается

+++ instrumental/src/main/java/by/bsuir/instrumental/packet/Packet.java
@@ -14,5 +14,5 @@ public class Packet implements Serializable {
     byte[] sourceId;
     byte[] targetId;
     short type;
-    int flags;
+    short flags;
 }
