+++ instrumental/src/main/java/by/bsuir/instrumental/packet/PacketFlags.java
@@ -5,8 +5,8 @@ import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 public enum PacketFlags {
-    ACK(0b1),
-    ABORT(0b10);
+    ACK((short) 0b1),
+    ABORT((short) 0b10);
     @Getter
-    public final int flagValue;
+    public final short flagValue;
 }
